+++ server/src-lib/Hasura/Server/Rest.hs
@@ -51,45 +51,23 @@ alignVars defVars parseVars =
     (M.fromList (map (\v -> (G._vdName v, v)) defVars))
     (M.fromList (mapMaybe (\(k, v) -> (,v) <$> G.mkName k) parseVars))
 
+-- | `resolveVar` is responsible for decoding variables sent via REST request.
+-- These can either be via body (represented by Right) or via query-param or URL param (represented by Left).
+-- A variable can be expected, unexpected, or missing (represented by These, This, and That).
 resolveVar :: G.Name -> These G.VariableDefinition (Either Text J.Value) -> Either Text (Maybe Value)
-resolveVar _ (This _expectedVar) = Right Nothing
-resolveVar varName (That _providedVar) = Left $ "Unexpected variable " <> toTxt @G.Name varName
-resolveVar varName (These expectedVar providedVar) =
-  -- TODO: See CustomTypes.hs for SCALAR types
+resolveVar _ (This _expectedVar) = Right Nothing -- If a variable is expected but missing, assign a missing value `Nothing` to it for resolution in query execution. This allows Null defaulting.
+resolveVar varName (That _providedVar) = Left $ "Unexpected variable " <> toTxt @G.Name varName -- If a variable is unexpected but present, throw an error.
+resolveVar _varName (These _expectedVar (Right bodyVar)) = Right (Just bodyVar) -- Variables sent via body can be passed through to execution without parsing.
+resolveVar varName (These expectedVar (Left l)) =
   case G._vdType expectedVar of
-    G.TypeNamed (G.Nullability nullable) typeName -> case providedVar of
-      Right r -> Right (Just r)
-      Left l
-        | typeName == stringScalar -> Right $ Just $ J.String l -- "String" -- Note: Strings don't need to be decoded since the format already matches.
-        | otherwise ->
-          case (J.decodeStrict (T.encodeUtf8 l), nullable) of
-            (Just J.Null, True) -> pure Nothing
-            (decoded, _)
-              | typeName == boolScalar && T.null l -> Right $ Just $ J.Bool True -- Key present but value missing for bools defaults to True.
-              | typeName == G._UUID -> Right $ Just $ J.String l
-              | typeName == G._uuid -> Right $ Just $ J.String l
-              | typeName == idScalar -> Right $ Just $ J.String l -- "ID" -- Note: Console doesn't expose this as a column type.
-              | otherwise -> case decoded of
-                (Just J.Null) -> Left $ "Null or missing value for non-nullable variable: " <> G.unName varName
-                (Just x@(J.Bool _))
-                  | typeName == boolScalar -> pure $ Just x -- "Boolean"
-                  | typeName == G._Bool -> pure $ Just x
-                  | otherwise -> Left $ "Expected " <> toTxt typeName <> " for variable " <> G.unName varName <> " got Bool"
-                (Just x@(J.Number _))
-                  | typeName == intScalar -> pure $ Just x -- "Int"
-                  | typeName == floatScalar -> pure $ Just x -- "Float"
-                  | typeName == G._Number -> pure $ Just x
-                  | typeName == G._Double -> pure $ Just x
-                  | typeName == G._float8 -> pure $ Just x
-                  | typeName == G._numeric -> pure $ Just x
-                  | otherwise -> Left $ "Expected " <> toTxt typeName <> " for variable " <> G.unName varName <> " got Number"
-                _ -> Left ("Type of URL parameter for variable " <> G.unName varName <> " not supported - Consider putting it in the request body: " <> tshow l)
-    -- TODO: This is a fallthrough case and is still required
-    --       but we can move checks for template variables being
-    --       scalars into the schema-cache construction.
-    G.TypeList _ _ -> case providedVar of
-      Right r -> Right (Just r)
-      Left l -> Left ("The variable type for the expected variable " <> toTxt @G.Name varName <> ", with value " <> tshow l <> " is not supported.")
+    G.TypeList _ _ -> Left $ "List variables are not currently supported in URL or Query parameters. (Variable " <> toTxt @G.Name varName <> ", with value " <> tshow l <> ")"
+    G.TypeNamed (G.Nullability nullable) typeName
+      | typeName == boolScalar && T.null l -> Right $ Just $ J.Bool True -- Booleans indicated true by a standalone key.
+      | nullable && T.null l -> Right Nothing -- Missing value, but nullable variable sets value to null.
+      | otherwise -> case J.decodeStrict (T.encodeUtf8 l) of -- We special case parsing of bools and numbers and pass the rest through as literal strings.
+        Just v@(J.Bool _) | typeName `elem` [G._Bool, boolScalar] -> Right $ Just v
+        Just v@(J.Number _) | typeName `elem` [intScalar, floatScalar, G._Number, G._Double, G._float8, G._numeric] -> Right $ Just v
+        _ -> Right $ Just $ J.String l
 
 mkPassthroughRequest :: EndpointMetadata GQLQueryWithText -> VariableValues -> GQLReq GQLQueryText
 mkPassthroughRequest queryx resolvedVariables =